Restructure as an Anagular-managed repo.
Because the repo was not structured to use `ng` to manage it, it was impossible to run `ng update`.  Rather than take the time to understand
the esoteric build process that was in place, I decided to update the repo to use `ng` to manage it.  I took the following steps:

First, I built a standalone repo with the correct structure.
- Created a standalone Angular repo on version 12, using the directions at https://v12.angular.io/guide/file-structure#multiple-projects.
- Created a library project for `angular-instantsearch`.
- Created applications for `e-commerce`, `media`, and `server-side-rendering`.

Second, I merged the missing structural elements into this repo.
- Copied necessary structure and support files from the standalone repo into this one.
- Moved the projects into their correct subfolders.

Finally, I removed unnecessary files that were linked to the previous configuration.
- Babel config files
- Rollup scripts

The vast majority of changes are simply renames (moves) of files to their new locations.  The most important addition is `angular.json`,
which `ng` uses to manage the projects.

Because the unit tests were written in Jest and `ng` does not play well with Jest by default, I only concerned myself with the proper
compilation of the code and not running the tests.  (We'll get to the tests later.)
- Added missing `@Directive` decorator to several classes
- Updated import paths
- Fix the public exports; Angular requires the components to be public
- Use `$any` in several templates to avoid compilation warning; not ideal!
- Fix CSS imports

Webpack errors can occur during the build, if so, attempt the following from the command-line:
export NODE_OPTIONS=--openssl-legacy-provider