Update to Angular 17.
Ran `ng update @angular/core@17 @angular/cli@17 --force`.
- `--force` was necessary because of peer dependency conflicts

Command completed successfully.

Updated `e-commerce` to wrap the `searchParameters` in `$any`.  The type is `PlainSearchParamters`, but it's clear that the code is looking
for `SearchParameters`.

Updated `server-side-rendering` to add injection tokens for the request.  This didn't migrate and I am unsure if this is the correct path
forward.  Will need to revisit.