Add Nx-managed ESLint back into the project.
This project was originally configured with ESLint, but we temporarily dropped it to become and Angular-managed repo.  Now that we're an
Nx-managed repo, we're able to add this back more easily.

Because we converted an Angular repo, this did not add handle the normal Nx construction of adding ESLint.  I had to use internal Nx
schematics to fix the ESLint.

Commands that were run:
- `nx g @nx/angular:add-linting --project-name=angular-instantsearch --project-root=libs --prefix=ais`
- `nx g @nx/angular:add-linting --project-name=e-commerce --project-root=apps --prefix=app`
- `nx g @nx/angular:add-linting --project-name=media --project-root=apps --prefix=app`
- `nx g @nx/angular:add-linting --project-name=server-side-rendering --project-root=apps --prefix=app`

Fixed all of the linting issues.  I had to create some exception cases, especially in `instantsearch.ts`.  We should revisit these in a clean
up pass.

Note: instead of fixing the warnings in some test files, I removed what I consider unnecessary tests.  There's no reason to create a test to
test tree-shaking.